requirements.txt
streamlit
google-generativeai
youtube-transcript-api
langchain
numpy
